moved ListGroupsForUser API view to separate file
