+++ poem/Poem/api/internal_views/app.py
@@ -1,0 +1,23 @@
+from Poem.api.internal_views.users import get_all_groups, get_groups_for_user
+
+from rest_framework.authentication import SessionAuthentication
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+
+class ListGroupsForUser(APIView):
+    authentication_classes = (SessionAuthentication,)
+
+    def get(self, request, group=None):
+        user = request.user
+
+        if user.is_superuser:
+            results = get_all_groups()
+
+        else:
+            results = get_groups_for_user(user)
+
+        if group:
+            return Response(results[group.lower()])
+        else:
+            return Response({'result': results})
